[kbn/optimizer] only build xpack examples when building xpack plugins (#78656)
