+++ x-pack/examples/alerting_example/README.md
