+++ x-pack/examples/alerting_example/common/constants.ts
@@ -1,0 +1,21 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+
+export const ALERTING_EXAMPLE_APP_ID = 'AlertingExample';
+
+// always firing
+export const DEFAULT_INSTANCES_TO_GENERATE = 5;
+
+// Astros
+export enum Craft {
+  OuterSpace = 'Outer Space',
+  ISS = 'ISS',
+}
+export enum Operator {
+  AreAbove = 'Are above',
+  AreBelow = 'Are below',
+  AreExactly = 'Are exactly',
+}
