+++ x-pack/examples/alerting_example/kibana.json
