+++ x-pack/examples/alerting_example/public/alert_types/astros.tsx
@@ -1,20 +1,7 @@
 /*
- * Licensed to Elasticsearch B.V. under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch B.V. licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
  */
 
 import React, { useState, useEffect, Fragment } from 'react';
@@ -32,9 +19,9 @@ import {
 import { i18n } from '@kbn/i18n';
 import { flatten } from 'lodash';
 import { ALERTING_EXAMPLE_APP_ID, Craft, Operator } from '../../common/constants';
-import { SanitizedAlert } from '../../../../x-pack/plugins/alerts/common';
-import { PluginSetupContract as AlertingSetup } from '../../../../x-pack/plugins/alerts/public';
-import { AlertTypeModel } from '../../../../x-pack/plugins/triggers_actions_ui/public';
+import { SanitizedAlert } from '../../../../plugins/alerts/common';
+import { PluginSetupContract as AlertingSetup } from '../../../../plugins/alerts/public';
+import { AlertTypeModel } from '../../../../plugins/triggers_actions_ui/public';
 
 export function registerNavigation(alerts: AlertingSetup) {
   alerts.registerNavigation(
