+++ x-pack/examples/alerting_example/server/index.ts
@@ -1,0 +1,10 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+
+import { PluginInitializer } from 'kibana/server';
+import { AlertingExamplePlugin } from './plugin';
+
+export const plugin: PluginInitializer<void, void> = () => new AlertingExamplePlugin();
