+++ x-pack/examples/alerting_example/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../../tsconfig.base.json",
+  "extends": "../../../tsconfig.base.json",
   "compilerOptions": {
     "outDir": "./target"
   },
@@ -9,10 +9,10 @@
     "public/**/*.tsx",
     "server/**/*.ts",
     "common/**/*.ts",
-    "../../typings/**/*",
+    "../../../typings/**/*",
   ],
   "exclude": [],
   "references": [
-    { "path": "../../src/core/tsconfig.json" }
+    { "path": "../../../src/core/tsconfig.json" }
   ]
 }
