+++ x-pack/plugins/translations/translations/zh-CN.json
@@ -18227,7 +18227,6 @@
     "xpack.triggersActionsUI.typeRegistry.register.duplicateObjectTypeErrorMessage": "已注册对象类型“{id}”。",
     "xpack.uiActionsEnhanced.components.actionWizard.changeButton": "更改",
     "xpack.uiActionsEnhanced.components.actionWizard.insufficientLicenseLevelTooltip": "许可证级别不够",
-    "xpack.uiActionsEnhanced.components.DiscoverDrilldownConfig.chooseIndexPattern": "选择目标索引模式",
     "xpack.uiActionsEnhanced.customizePanelTimeRange.modal.addToPanelButtonTitle": "添加到面板",
     "xpack.uiActionsEnhanced.customizePanelTimeRange.modal.cancelButtonTitle": "取消",
     "xpack.uiActionsEnhanced.customizePanelTimeRange.modal.optionsMenuForm.panelTitleFormRowLabel": "时间范围",
@@ -18235,7 +18234,6 @@
     "xpack.uiActionsEnhanced.customizePanelTimeRange.modal.updatePanelTimeRangeButtonTitle": "更新",
     "xpack.uiActionsEnhanced.customizeTimeRange.modal.headerTitle": "定制面板时间范围",
     "xpack.uiActionsEnhanced.customizeTimeRangeMenuItem.displayName": "定制时间范围",
-    "xpack.uiActionsEnhanced.drilldown.goToDiscover": "前往 Discover（示例）",
     "xpack.uiActionsEnhanced.drilldowns.components.DrilldownHelloBar.helpText": "向下钻取允许您定义与面板交互的新行为。您可以添加多个操作并覆盖默认筛选。",
     "xpack.uiActionsEnhanced.drilldowns.components.DrilldownHelloBar.hideHelpButtonLabel": "隐藏",
     "xpack.uiActionsEnhanced.drilldowns.components.DrilldownHelloBar.viewDocsLinkLabel": "查看文档",
